debug: typed enums on the write side, two silent wrong answers, and one measured refusal (08 Track H)
H1 -- F-32.  The address-space enumeration was a V_ENUM on the read side
(bp.space) and a V_STRING plus a hand-rolled strcmp on the two method
ARGUMENTS, in the same file.  So reads were typed and writes were not: an
unrecognised string silently meant "logical", nothing could complete the two
values, and object-model.md lists enum membership as something bodies must not
re-check.  One table, declared once, used by the getter and both arg slots;
the strcmps are gone and lp_def_space becomes a V_ENUM so the default agrees
with what validate_slot produces for an explicit "logical".

H3 -- two bugs that give a debugger a silently wrong answer, which is the
worst failure mode a debugging tool has.

F-41: the per-page logpoint refcounts saturated at 255 on install while every
uninstall decremented, so the counter stopped being a refcount the moment it
saturated.  With 300 logpoints on one page, installs 256..300 did not
increment, removing 255 of them drove the count to 0, rebuild_soa_page()
restored the direct mapping, and the 45 SURVIVING logpoints on that page
stopped firing -- making you conclude the guest never touched the address.
uint16_t now, one extra byte per 4 KB of address space, with an assert on
saturation so the successor is loud instead of silent.

F-40: set_memory_logpoint used malloc where set_logpoint uses calloc and then
assigned 14 of 16 fields by hand, leaving start_phys/end_phys as heap garbage.
Inert today only because the single reader is guarded by kind == LP_KIND_PC --
a coincidence of control flow, not a stated invariant, and H3's own widening
is the kind of change that would break it.  calloc plus the explicit no-range
sentinel, matching set_logpoint.

H4 -- F-53 and F-54.  The completer filtered candidates case-INsensitively
while the resolver matches with strcmp, so `MACH<Tab>` offered `machine`,
which then failed to resolve -- and push_match's own dedup two lines below was
already case-sensitive, so the function disagreed with itself.  And a dropped
candidate (the 2 KB pool filled, or the item table did) was indistinguishable
from a complete set: struct completion gains `truncated`, and its declaration
now states the pointer lifetime that was previously true only because
shell_meta_complete_provider happens to copy immediately.

H2 -- F-42, MEASURED and deliberately not restructured.

The finding says a `message=` template is re-parsed on every fire and a write
logpoint on a hot page is "effectively unusable".  The first half is true.
The second is not, and the work order's rule for this unit is to measure
before believing the severity.  A Plus, a write logpoint over 0x0000-0xFFFF,
20M cycles, 71,650 fires in every run, output discarded, best of three:

    no message=          455 ms
    plain message=       468 ms   (+13 ms; no ${} so it short-circuits)
    two-expression ${}=  529 ms   (+61 ms over plain)

0.85 us per fire, about 13% -- on top of a path already costing 6.4 us per
fire for the slow-path routing the logpoint itself forces.  The template is
not what makes a hot-page logpoint expensive.  Pre-splitting it into a chunk
list would add cached state and its invalidation to recover a small fraction
of an already-slow debugging path, so it is not done, and the number to beat
is recorded in the code rather than in a commit message nobody greps.

Verified: unit tier 882 pass / 0 fail; `make headless` and `make` both build,
and the warning set for memory.c, debug.c and cmd_complete.c is SMALLER than
main's (7 unique before, 5 after -- widening the refcounts removed two
sign-compare warnings); integration rows object-logpoint (which exercises a
message template through the path H2 measured), debug and shell-v2 pass;
clang-format-18 clean.

+++ src/core/memory/memory.h
@@ -349,13 +349,13 @@ bool memory_debug_write_uint32(uint32_t addr, uint32_t value);
 // Per-page memory-logpoint reference count.  Non-zero entries indicate pages
 // whose SoA fast-path must stay at 0 (force slow path). Allocated alongside the
 // page tables in memory_map_init.
-extern uint8_t *g_mem_logpoint_page_count;
+extern uint16_t *g_mem_logpoint_page_count;
 
 // Per-physical-page memory-logpoint reference count.  Non-zero entries mean
 // "any logical alias mapping to this physical page must stay on the slow
 // path so the logpoint fires regardless of which alias the CPU uses."
 // Indexed by physical page number.  mmu_fill_soa_entry consults both arrays.
-extern uint8_t *g_mem_logpoint_phys_page_count;
+extern uint16_t *g_mem_logpoint_phys_page_count;
 
 // Hook invoked by the slow path on logpoint pages.  is_write=true on writes.
 // Installed by debug.c.  NULL means no hook (skip check).
